Re-shuffle workspace structure
