+++ README.md
@@ -7,7 +7,7 @@ Design and Deploy Large Language Model Apps
 ## :book: Documentation Portal
 
 - [LLM apps Primer](https://docs.dust.tt/introduction)
-- [Platform overview](https://docs.dust.tt/overview)
+- [Developer Platform overview](https://docs.dust.tt/overview)
 - [Blocks documentation](https://docs.dust.tt/core-blocks)
 - [Getting started guide](https://docs.dust.tt/quickstart)
 - [Runs API Reference](https://docs.dust.tt/runs)
