+++ docs/src/pages/index.mdx
@@ -2,11 +2,11 @@ import { Guides } from "@/components/Guides";
 import { Resources } from "@/components/Resources";
 
 export const description =
-  "Learn everything there is to know about the Dust Platform and how to build LLM apps with it.";
+  "Learn everything there is to know about the Dust Developer Platform and how to build LLM apps with it.";
 
-# Dust Platform Documentation
+# Dust Developer Platform Documentation
 
-Use the Dust Platform to design and deploy Large Language Model apps. {{ className: 'lead' }}
+Use the Dust Developer Platform to design and deploy Large Language Model apps. {{ className: 'lead' }}
 
 <div className="not-prose mb-16 mt-6 flex gap-3">
   <Button href="/introduction" arrow="right" children="Introduction" />
