+++ docs/src/pages/overview.mdx
@@ -1,10 +1,10 @@
 export const description =
-  "An overview of the Dust Platform along with its associated design principles.";
+  "An overview of the Dust Developer Platform along with its associated design principles.";
 
-# Platform overview
+# Developer Platform overview
 
-An overview of the Dust Platform along with its associated design principles. You will learn how a
-Dust app is structured and how it is executed. {{ className: 'lead' }}
+An overview of the Dust Developer Platform along with its associated design principles. You will
+learn how a Dust app is structured and how it is executed. {{ className: 'lead' }}
 
 Dust is designed to provide a flexible framework to define and deploy [large language model
 apps](/introduction#large-language-model-apps) without having to write any execution code. It is
@@ -84,8 +84,8 @@ an app with various configurations. As an example, this can be used to evaluate 
 different models. To illustrate the distinction, the `llm` block's _prompt_ is part of its
 specification whereas the _model_ is part of its configuration.
 
-The Dust Platform UI does not explicitly expose the distinction. But you can arbitrarily change the
-configuration of each blocks of an app each time you call it by API.
+The Dust Developer Platform UI does not explicitly expose the distinction. But you can arbitrarily
+change the configuration of each blocks of an app each time you call it by API.
 
 ## Inputs
